+++ src/types.ts
@@ -95,10 +95,7 @@ export interface InfluxQuery extends DataQuery {
   timeEndColumn?: string;
   titleColumn?: string;
   name?: string;
-  matchAny?: boolean;
-  type?: string;
 
-  textEditor?: boolean;
   adhocFilters?: AdHocVariableFilter[];
 }